+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Scriptlike - Changelog
 
 (Dates below are YYYY/MM/DD)
 
+v0.7.1 - TBD
+-------------------
+- **Enhancement:** ```fail()``` now accepts an arbitrary list of args or any type, just like writeln,
+- **Enhancement:** Added [```enforceFail```](#), like std.exception.enforce, but for ```fail()```.
+
 v0.7.0 - 2015/04/02
 -------------------
 - **Enhancement:** [#14](https://github.com/Abscissa/scriptlike/issues/14): Added ```[interact](http://semitwist.com/scriptlike/interact.html)``` module for easy user-input prompts. (Thanks to [Jesse Phillips](https://github.com/JesseKPhillips))